+++ src/fwbuf.c
@@ -701,14 +701,14 @@ fwbuf_map(fwbuf_t *buf, int64_t pos, size_t *psize)
 	ULARGE_INTEGER FileOffset = { .QuadPart = pos - off };
 	buf->lpBaseAddress = MapViewOfFile(buf->hFileMappingObject,
 			FILE_MAP_WRITE, FileOffset.HighPart, FileOffset.LowPart,
-			off + size);
+			(SIZE_T)(off + size));
 	if (__unlikely(!buf->lpBaseAddress)) {
 		buf->dwErrCode = GetLastError();
 		goto error_MapViewOfFile;
 	}
 
 	if (psize)
-		*psize = size;
+		*psize = (size_t)size;
 
 	return (char *)buf->lpBaseAddress + off;
